requirements.txt
# Core dependencies
mysql-connector-python==8.2.0
googlemaps==4.10.0
requests==2.31.0

# GUI enhancements
tkinter-tooltip==2.0.0

# Database and ORM - Fixed for Python 3.13 compatibility
sqlalchemy==2.0.36
pymysql==1.1.0

# Configuration and environment
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Validation and security
bcrypt==4.2.0